+++ test/unit/range/views/CMakeLists.txt
@@ -21,4 +21,3 @@ seqan3_test(to_upper_test.cpp)
 seqan3_test(translate_join_test.cpp)
 seqan3_test(translate_test.cpp)
 seqan3_test(trim_quality_test.cpp)
-seqan3_test(interleave_test.cpp)
