+++ test/unit/utility/views/interleave_test.cpp
@@ -12,11 +12,11 @@
 
 #include <seqan3/alphabet/nucleotide/dna4.hpp>
 #include <seqan3/core/detail/debug_stream_alphabet.hpp>
-#include <seqan3/range/views/interleave.hpp>
 #include <seqan3/range/views/take.hpp>
 #include <seqan3/range/views/type_reduce.hpp>
 #include <seqan3/test/expect_range_eq.hpp>
 #include <seqan3/test/pretty_printing.hpp>
+#include <seqan3/utility/views/interleave.hpp>
 
 using seqan3::operator""_dna4;
